# backend/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Team tuner API documentation",
        "title": "TT API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/group": {
            "get": {
                "description": "Get groups",
                "tags": [
                    "groups"
                ],
                "summary": "Get groups",
                "operationId": "get-groups-by-group-run",
                "parameters": [
                    {
                        "type": "string",
                        "name": "groupRun",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Groups filtered by query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/out.GroupOut"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    }
                }
            }
        },
        "/group/download/{groupRunId}": {
            "get": {
                "description": "Download groups as CSV",
                "tags": [
                    "groups"
                ],
                "summary": "Download groups as CSV",
                "operationId": "download-groups-csv",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group run id",
                        "name": "groupRunId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    }
                }
            }
        },
        "/group/form": {
            "post": {
                "description": "Create groups for a course. Runs an algorithm to assign students to groups. After the algorithm finished the following are created: A group run and groups with students asiigned to them",
                "tags": [
                    "groups"
                ],
                "summary": "Create group",
                "operationId": "create-groups",
                "parameters": [
                    {
                        "description": "Create groups request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/in.PostFormGroupsIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/out.Error"
                        }
                    }
                }
            }
        },
        "/group/{id}": {
            "get": {
                "description": "Get group by id: their students and stats (bfi scores)",
                "tags": [
                    "groups"
                ],
                "summary": "Get group by id: their students and stats",
                "operationId": "get-group-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group run id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/out.DetailedGroupOut"
                        }
                    }
                }
            }
        },
...
        "/question": {
            "get": {
                "description": "Get questions",
                "tags": [
                    "questions"
                ],
                "summary": "Get questions",
                "parameters": [
                    {
                        "enum": [
                            "extroversion",
                            "agreeableness",
                            "conscientiousness",
                            "neuroticism",
                            "openness"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "QuestionTypesExtroversion",
                            "QuestionTypesAgreeableness",
                            "QuestionTypesConscientiousness",
                            "QuestionTypesNeuroticism",
                            "QuestionTypesOpenness"
                        ],
                        "name": "typeOfQuestion",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/out.QuestionerOut"
                        }
                    }
                }
            }
        },
        "/question/answer": {
            "post": {
                "description": "Post questions answer",
                "tags": [
                    "questions"
                ],
                "summary": "Post questions answer",
                "parameters": [
                    {
                        "description": "Post questions answer body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/in.PostQuestionAnswerIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/out.QuestionerOut"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "in.PostFormGroupsIn": {
            "type": "object",
            "required": [
                "courseId",
                "groupSizes",
                "iterations",
                "temperature",
                "weightConscientiousness",
                "weightExtraversion",
                "weightNeuroticism",
                "weightSatisfaction"
            ],
            "properties": {
                "courseId": {
                    "description": "To which course the groups belong",
                    "type": "string"
                },
                "groupSizes": {
                    "description": "Array length must be at least 4\nEach element must be greater than 0 and less than 10",
                    "type": "array",
                    "minItems": 4,
                    "items": {
                        "type": "integer"
                    }
                },
                "iterations": {
                    "description": "General settings for SA",
                    "type": "integer"
                },
                "temperature": {
                    "type": "number"
                },
                "weightConscientiousness": {
                    "type": "number"
                },
                "weightExtraversion": {
                    "type": "number"
                },
                "weightNeuroticism": {
                    "type": "number"
                },
                "weightSatisfaction": {
                    "description": "Weights for SA",
                    "type": "number"
                }
            }
        },
        "in.PostQuestionAnswerIn": {
            "type": "object",
            "required": [
                "answerID",
                "value"
            ],
            "properties": {
                "answerID": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
...
        "out.BigFiveBoxPlot": {
            "type": "object",
            "required": [
                "agreeableness",
                "conscientious",
                "extraversion",
                "neuroticism",
                "openness"
            ],
            "properties": {
                "agreeableness": {
                    "$ref": "#/definitions/out.BoxPlotItem"
                },
                "conscientious": {
                    "$ref": "#/definitions/out.BoxPlotItem"
                },
                "extraversion": {
                    "$ref": "#/definitions/out.BoxPlotItem"
                },
                "neuroticism": {
                    "$ref": "#/definitions/out.BoxPlotItem"
                },
                "openness": {
                    "$ref": "#/definitions/out.BoxPlotItem"
                }
            }
        },
        "out.BoxPlotItem": {
            "type": "object",
            "required": [
                "dataPoints",
                "max",
                "mean",
                "min",
                "q1",
                "q3"
            ],
            "properties": {
                "dataPoints": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "max": {
                    "type": "number"
                },
                "mean": {
                    "type": "number"
                },
                "min": {
                    "type": "number"
                },
                "q1": {
                    "type": "number"
                },
                "q3": {
                    "type": "number"
                }
            }
        },
...
        "out.DetailedGroupOut": {
            "type": "object",
            "required": [
                "bigFiveBoxPlot",
                "courseCode",
                "courseColour",
                "courseName",
                "createdAt",
                "id",
                "name",
                "students",
                "updateAt"
            ],
            "properties": {
                "bigFiveBoxPlot": {
                    "$ref": "#/definitions/out.BigFiveBoxPlot"
                },
                "courseCode": {
                    "type": "string"
                },
                "courseColour": {
                    "type": "string"
                },
                "courseName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "students": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/out.StudentOut"
                    }
                },
                "updateAt": {
                    "type": "string"
                }
            }
        },
        "out.Error": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "out.GroupOut": {
            "type": "object",
            "required": [
                "courseCode",
                "courseColour",
                "courseName",
                "createdAt",
                "id",
                "name",
                "students",
                "updateAt"
            ],
            "properties": {
                "courseCode": {
                    "type": "string"
                },
                "courseColour": {
                    "type": "string"
                },
                "courseName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "students": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/out.StudentOut"
                    }
                },
                "updateAt": {
                    "type": "string"
                }
            }
        },
...
        "out.QuesitonOutType": {
            "type": "string",
            "enum": [
                "extraversion",
                "agreeableness",
                "conscientious",
                "neuroticism",
                "openness"
            ],
            "x-enum-varnames": [
                "QuestionTypeExtraversion",
                "QuesitonOutTypeAgreeableness",
                "QuesitonOutTypeConscientious",
                "QuesitonOutTypeNeuroticism",
                "QuesitonOutTypeOpenness"
            ]
        },
        "out.QuestioneItemOut": {
            "type": "object",
            "required": [
                "answer_id",
                "question",
                "question_id",
                "question_type"
            ],
            "properties": {
                "answer_id": {
                    "type": "string"
                },
                "answer_value": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "question_id": {
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/out.QuesitonOutType"
                }
            }
        },
...
        "out.QuestionerOut": {
            "type": "object",
            "required": [
                "description",
                "questions",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/out.QuestioneItemOut"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "out.StudentOut": {
            "type": "object",
            "required": [
                "completedQuestioner",
                "createdAt",
                "githubUsername",
                "id",
                "updatedAt"
            ],
            "properties": {
                "agreeableness": {
                    "type": "number"
                },
                "completedQuestioner": {
                    "type": "boolean"
                },
                "completedQuestionerAt": {
                    "type": "string"
                },
                "concientiousness": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "extraversion": {
                    "type": "number"
                },
                "githubUsername": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "neuroticism": {
                    "type": "number"
                },
                "openness": {
                    "type": "number"
                },
                "univeresityId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}